pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repo2-inference-system"
version = "0.1.0"
description = "Base monorepo for a self-hosted inference and RAG platform"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1",
  "httpx>=0.27,<1",
  "pydantic>=2.8,<3",
  "PyYAML>=6,<7",
  "uvicorn>=0.30,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
]

[project.scripts]
repo2ctl = "repo2ctl.cli:main"

[tool.pytest.ini_options]
pythonpath = [".", "clients/python"]
testpaths = ["tests", "services", "clients"]

[tool.setuptools]
include-package-data = true
